requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
asyncio==3.4.3
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
gotrue==2.10.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
h5py==3.13.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.29.1
hyperframe==6.0.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langdetect==1.0.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mpmath==1.3.0
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.1.3
onnx==1.17.0
openai==1.58.1
opencv-python==4.10.0.84
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
postgrest==0.18.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.1
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-doctr==0.11.0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.17
pytorch-triton-rocm==3.2.0
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.1
realtime==2.0.6
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.2
selenium==4.27.1
Send2Trash==1.8.3
shapely==2.0.7
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.2
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
sympy==1.13.1
terminado==0.18.1
tinycss2==1.4.0
torch==2.6.0+rocm6.2.4
torchaudio==2.6.0+rocm6.2.4
torchvision==0.21.0+rocm6.2.4
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
twilio==9.4.1
typer==0.13.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
widgetsnbextension==4.0.13
wsproto==1.2.0
yarl==1.17.1
